+++ lib/nikfemm/src/algebra/csr.hpp
@@ -12,7 +12,7 @@ namespace nikfemm {
     struct MatCSR {
         uint32_t* IA;
         uint32_t* JA;
-        double* A;
+        float* A;
 
         uint32_t nnz;
         uint32_t m, n;  // I, J  // rows,  columns
@@ -24,14 +24,14 @@ namespace nikfemm {
         void print();
         void write_to_file(const char *filename);
 
-        double operator()(uint32_t i, uint32_t j) const;
+        float operator()(uint32_t i, uint32_t j) const;
         
         CV operator*(const CV& cv) const;
         
         CV getInverseDiagonal() const;
 
-        void conjugateGradientSolve(CV& b, CV& x0, double maxError, uint32_t maxIterations);
-        void preconditionedConjugateGradientSolve(CV& b, CV& x0, double maxError, uint32_t maxIterations);
+        void conjugateGradientSolve(CV& b, CV& x0, float maxError, uint32_t maxIterations);
+        void preconditionedConjugateGradientSolve(CV& b, CV& x0, float maxError, uint32_t maxIterations);
     };
 }
